lambda/auth/requirements.txt
bcrypt==4.1.2
PyJWT==2.8.0
